+++ RedPandaIDE/translations/RedPandaIDE_zh_TW.ts
@@ -6115,10 +6115,6 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>Check ISO C/C++/C++0x conformance (-pedantic)</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
         <source>Only check the code for syntax errors (-fsyntax-only)</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6280,6 +6276,18 @@
     </message>
     <message>
         <source>Lua files</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Check for stack smashing attacks (-fstack-protector)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Check ISO C/C++ conformance (-pedantic)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Enable Sanitizer (-fsanitize=)</source>
         <translation type="unfinished"></translation>
     </message>
 </context>